requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
boto3==1.33.4
botocore==1.33.4
certifi==2023.11.17
charset-normalizer==3.3.2
decorator==4.4.2
distro==1.8.0
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
imageio==2.33.0
imageio-ffmpeg==0.4.9
jmespath==1.0.1
moviepy==1.0.3
multidict==6.0.4
numpy==1.26.2
openai==1.2.0
Pillow==10.1.0
proglog==0.1.10
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
python-dateutil==2.8.2
requests==2.31.0
s3transfer==0.8.2
six==1.16.0
sniffio==1.3.0
SpeechRecognition==3.10.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
yarl==1.9.3
python-dotenv
kafka-python
pymongo